Longer commit

+++ src/Main.java
@@ -1,6 +1,6 @@
 public class Main {
     public static void main(String[] args) {
-        int [] s = {100, 200, 300};
+        long [] s = {100, 200, 300};
 
         SalesManager sm = new SalesManager(s);
         System.out.println("Максимальная продажа " + sm.max());
